# fd-microservice/openapi_spec.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "OpenAPI definition",
                 "version":  "v0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Generated server url"
                    }
                ],
    "paths":  {
...
                  "/api/fd/calculator/simulate":  {
                                                      "post":  {
                                                                   "tags":  [
                                                                                "fd-calculator-controller"
                                                                            ],
                                                                   "operationId":  "simulate",
                                                                   "requestBody":  {
                                                                                       "content":  {
                                                                                                       "application/json":  {
                                                                                                                                "schema":  {
                                                                                                                                               "$ref":  "#/components/schemas/FdCalculateRequest"
                                                                                                                                           }
                                                                                                                            }
                                                                                                   },
                                                                                       "required":  true
                                                                                   },
                                                                   "responses":  {
                                                                                     "404":  {
                                                                                                 "description":  "Not Found",
                                                                                                 "content":  {
                                                                                                                 "*/*":  {
                                                                                                                             "schema":  {
                                                                                                                                            "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                        }
                                                                                                                         }
                                                                                                             }
                                                                                             },
                                                                                     "400":  {
                                                                                                 "description":  "Bad Request",
                                                                                                 "content":  {
                                                                                                                 "*/*":  {
                                                                                                                             "schema":  {
                                                                                                                                            "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                        }
                                                                                                                         }
                                                                                                             }
                                                                                             },
                                                                                     "500":  {
                                                                                                 "description":  "Internal Server Error",
                                                                                                 "content":  {
                                                                                                                 "*/*":  {
                                                                                                                             "schema":  {
                                                                                                                                            "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                        }
                                                                                                                         }
                                                                                                             }
                                                                                             },
                                                                                     "403":  {
                                                                                                 "description":  "Forbidden",
                                                                                                 "content":  {
                                                                                                                 "*/*":  {
                                                                                                                             "schema":  {
                                                                                                                                            "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                        }
                                                                                                                         }
                                                                                                             }
                                                                                             },
                                                                                     "200":  {
                                                                                                 "description":  "OK",
                                                                                                 "content":  {
                                                                                                                 "*/*":  {
                                                                                                                             "schema":  {
                                                                                                                                            "$ref":  "#/components/schemas/FdCalculateResponse"
                                                                                                                                        }
                                                                                                                         }
                                                                                                             }
                                                                                             }
                                                                                 }
                                                               }
                                                  },
...
              },
    "components":  {
                       "schemas":  {
                                       "ApiResponse":  {
                                                           "type":  "object",
                                                           "properties":  {
                                                                              "success":  {
                                                                                              "type":  "boolean"
                                                                                          },
                                                                              "message":  {
                                                                                              "type":  "string"
                                                                                          },
                                                                              "data":  {
                                                                                           "type":  "object"
                                                                                       },
                                                                              "timestamp":  {
                                                                                                "type":  "string",
                                                                                                "format":  "date-time"
                                                                                            }
                                                                          }
                                                       },
...
                                       "FdCalculateRequest":  {
                                                                  "required":  [
                                                                                   "baseRate",
                                                                                   "principal",
                                                                                   "termMonths"
                                                                               ],
                                                                  "type":  "object",
                                                                  "properties":  {
                                                                                     "principal":  {
                                                                                                       "minimum":  1000,
                                                                                                       "exclusiveMinimum":  false,
                                                                                                       "type":  "number"
                                                                                                   },
                                                                                     "termMonths":  {
                                                                                                        "minimum":  1,
                                                                                                        "type":  "integer",
                                                                                                        "format":  "int32"
                                                                                                    },
                                                                                     "baseRate":  {
                                                                                                      "type":  "number"
                                                                                                  },
                                                                                     "compoundingFrequency":  {
                                                                                                                  "type":  "string"
                                                                                                              },
                                                                                     "categories":  {
                                                                                                        "type":  "array",
                                                                                                        "items":  {
                                                                                                                      "type":  "string"
                                                                                                                  }
                                                                                                    },
                                                                                     "calculationType":  {
                                                                                                             "type":  "string"
                                                                                                         }
                                                                                 }
                                                              },
                                       "FdCalculateResponse":  {
                                                                   "type":  "object",
                                                                   "properties":  {
                                                                                      "principal":  {
                                                                                                        "type":  "number"
                                                                                                    },
                                                                                      "effectiveRate":  {
                                                                                                            "type":  "number"
                                                                                                        },
                                                                                      "tenureMonths":  {
                                                                                                           "type":  "integer",
                                                                                                           "format":  "int32"
                                                                                                       },
                                                                                      "compoundingFrequency":  {
                                                                                                                   "type":  "string"
                                                                                                               },
                                                                                      "calculationType":  {
                                                                                                              "type":  "string"
                                                                                                          },
                                                                                      "maturityAmount":  {
                                                                                                             "type":  "number"
                                                                                                         },
                                                                                      "interestEarned":  {
                                                                                                             "type":  "number"
                                                                                                         },
                                                                                      "categoryAddons":  {
                                                                                                             "type":  "object",
                                                                                                             "additionalProperties":  {
                                                                                                                                          "type":  "number"
                                                                                                                                      }
                                                                                                         },
                                                                                      "totalAddon":  {
                                                                                                         "type":  "number"
                                                                                                     }
                                                                                  }
                                                               },
...
                                   }
                   }
}